bugfix with most_listened genre and floats

+++ app/database/migrations/2014_09_07_164833_diary_entry_spotify.php
@@ -23,14 +23,14 @@ class DiaryEntrySpotify extends Migration {
 						
 			$table->integer('listening_duration')->default(0);
 			
-			$table->integer('listened_to_browse_top_lists')->default(0);
-			$table->integer('listened_to_browse_news')->default(0);
-			$table->integer('listened_to_browse_moods')->default(0);
-			$table->integer('listened_to_browse_discover')->default(0);
-			$table->integer('listened_to_own_playlist')->default(0);
-			$table->integer('listened_to_other_playlist')->default(0);
-			$table->integer('listened_to_searched_songs')->default(0);
-			$table->integer('listened_to_radio')->default(0);
+			$table->float('listened_to_browse_top_lists')->default(0);
+			$table->float('listened_to_browse_news')->default(0);
+			$table->float('listened_to_browse_moods')->default(0);
+			$table->float('listened_to_browse_discover')->default(0);
+			$table->float('listened_to_own_playlist')->default(0);
+			$table->float('listened_to_other_playlist')->default(0);
+			$table->float('listened_to_searched_songs')->default(0);
+			$table->float('listened_to_radio')->default(0);
 			
 			$table->integer('recommendations_quality')->default(0);
 			$table->integer('recommendations_comparison')->default(0);
